Added cache tests to Movie Details ViewModel
